refactor: create graphql definitions mod
